combine two logically related classes
